Add Room

+++ src/main/java/edu/uniandes/hotelandes/hotel/room/type/RoomType.java
@@ -1,7 +1,4 @@
 package edu.uniandes.hotelandes.hotel.room.type;
 
 public record RoomType(
-    Short id,
-    Double costPerNight,
-    Short suggestedMinimumCapacity,
-    Short suggestedMaximumCapacity) {}
+    Byte id, Double costPerNight, Short suggestedMinimumCapacity, Short suggestedMaximumCapacity) {}
